COM-2975-2 mutualise getNextStep between TrainerCourses and LearnerCourses
