+++ src/screens/courses/list/LearnerCourses/index.tsx
@@ -6,24 +6,22 @@ import { connect } from 'react-redux';
 import { useIsFocused, CompositeScreenProps } from '@react-navigation/native';
 import { StackScreenProps } from '@react-navigation/stack';
 import get from 'lodash/get';
-import groupBy from 'lodash/groupBy';
 import Courses from '../../../../api/courses';
 import SubPrograms from '../../../../api/subPrograms';
 import NextStepCell from '../../../../components/steps/NextStepCell';
 import ProgramCell from '../../../../components/ProgramCell';
 import CoursesSection, { EVENT_SECTION } from '../../../../components/CoursesSection';
-import CompaniDate from '../../../../core/helpers/dates/companiDates';
-import { ascendingSort } from '../../../../core/helpers/dates/utils';
 import { getLoggedUserId } from '../../../../store/main/selectors';
 import CoursesActions from '../../../../store/courses/actions';
 import commonStyles from '../../../../styles/common';
 import { RootBottomTabParamList, RootStackParamList } from '../../../../types/NavigationType';
-import { CourseType, BlendedCourseType, SubProgramType } from '../../../../types/CourseTypes';
+import { SubProgramType } from '../../../../types/CourseTypes';
 import { NextSlotsStepType } from '../../../../types/StepTypes';
 import { ActionWithoutPayloadType } from '../../../../types/store/StoreType';
 import { getCourseProgress, getTheoreticalHours } from '../../../../core/helpers/utils';
 import { E_LEARNING } from '../../../../core/data/constants';
 import styles from '../styles';
+import { getNextSteps } from '../helper';
 import LearnerEmptyState from '../LearnerEmptyState';
 
 interface LearnerCoursesProps extends CompositeScreenProps<
@@ -33,40 +31,6 @@ StackScreenProps<RootStackParamList>
   setIsCourse: (value: boolean) => void,
   loggedUserId: string | null,
 }
-
-const formatCourseStep = (stepId: string, course: CourseType, stepSlots): NextSlotsStepType => {
-  const courseSteps = get(course, 'subProgram.steps') || [];
-  const nextSlots = stepSlots[stepId].filter(slot => CompaniDate().isSameOrBefore(slot.endDate));
-  const slotsSorted = stepSlots[stepId].sort(ascendingSort('endDate'));
-  const stepIndex = courseSteps.map(step => step._id).indexOf(stepId);
-
-  return {
-    name: get(course, 'subProgram.program.name'),
-    stepIndex,
-    firstSlot: nextSlots[0].endDate,
-    type: nextSlots[0].step.type,
-    slots: slotsSorted.map(s => s.endDate),
-    _id: slotsSorted[0]._id,
-    progress: courseSteps[stepIndex].progress,
-    courseId: course._id,
-  };
-};
-
-const formatNextSteps = (course: CourseType): NextSlotsStepType[] => {
-  if (course.subProgram.isStrictlyELearning) return [];
-
-  const { slots } = course as BlendedCourseType;
-  const stepSlots = groupBy(slots.filter(s => get(s, 'step._id')), s => s.step._id);
-
-  return Object.keys(stepSlots)
-    .filter(stepId => stepSlots[stepId].some(slot => CompaniDate().isSameOrBefore(slot.endDate)))
-    .map(stepId => formatCourseStep(stepId, course, stepSlots));
-};
-
-const getNextSteps = (courses: CourseType[]): NextSlotsStepType[] => courses.map(formatNextSteps)
-  .flat()
-  .filter(step => step.slots && step.slots.length)
-  .sort(ascendingSort('firstSlot'));
 
 const SET_COURSES = 'SET_COURSES';
 const RESET_COURSES = 'RESET_COURSES';
@@ -140,7 +104,7 @@ const LearnerCourses = ({ setIsCourse, navigation, loggedUserId }: LearnerCourse
     theoreticalHours={getTheoreticalHours(getElearningSteps(subProgram.steps))}
     program={get(subProgram, 'program') || {}} />;
 
-  const nextSteps = useMemo(() => getNextSteps(courses.onGoing), [courses.onGoing]);
+  const nextSteps: NextSlotsStepType[] = useMemo(() => getNextSteps(courses.onGoing), [courses.onGoing]);
 
   return (
     <SafeAreaView style={commonStyles.container} edges={['top']}>
