+++ src/types/StepTypes.tsx
@@ -15,7 +15,7 @@ export type StepType = ELearningStepType | LiveStepType;
 
 export type NextSlotsStepType = BaseStepType & {
   type: typeof ON_SITE | typeof REMOTE,
-  stepIndex: string,
+  stepIndex: number,
   firstSlot: Date,
   slots: Date[],
   courseId: string,
